Refactor: Move assessment dtos (#364)
* Migrate more questions to use GenericTextField (#360)

* move error text inside GenericTextField

* migrate birth year + postcode question

* migrate LevelOfIsolation text inputs

* migrate period question inputs

* fix ts-ignore

* add showError flag

* Refactor Height + Weight Questions (#362)

* refactor height question

* refactor weight question

* remove unused comment lines

* fix imports

* Move Assessment DTOs into core/assessment/dto

* Move Assessment DTOs into core/assessment/dto
