+++ src/core/assessment/dto/AssessmentInfosResponse.ts
@@ -1,0 +1,3 @@
+export type AssessmentResponse = {
+  id: string;
+};
