+++ src/features/assessment/DescribeSymptomsScreen.tsx
@@ -5,8 +5,8 @@ import Screen, { FieldWrapper, Header, ProgressBlock } from '@covid/components/S
 import { BrandedButton, ErrorText, HeaderText } from '@covid/components/Text';
 import { ValidatedTextInput } from '@covid/components/ValidatedTextInput';
 import { ValidationErrors } from '@covid/components/ValidationError';
+import { AssessmentInfosRequest } from '@covid/core/assessment/dto/AssessmentInfosRequest';
 import UserService from '@covid/core/user/UserService';
-import { AssessmentInfosRequest } from '@covid/core/user/dto/UserAPIContracts';
 import { cleanFloatVal } from '@covid/core/utils/number';
 import i18n from '@covid/locale/i18n';
 import { RouteProp } from '@react-navigation/native';
